pyproject.toml
[project]
name = "rng-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.0",
    "python-dotenv>=1.1.1",
]